absensi-wajah-siswa/requirements.txt
fastapi
uvicorn
python-multipart
Pillow
numpy
opencv-python
joblib
torch
facenet-pytorch
